requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
Pillow==10.1.0
django-environ==0.11.2
redis==5.0.1
django-redis==5.4.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==23.3
drf-spectacular==0.26.5
whitenoise==6.6.0
requests==2.31.0
yookassa==3.0.0
reportlab==4.0.7
resend==2.0.0
cryptography==42.0.5
